Move rbenv.rb contents to system.rb
